git-svn-id: file:///var/svn/claro-catalog@618 0db79848-0afd-4259-9b79-6da5ec97fbb6

+++ trunk/catalog-model/source/claro/catalog/impl/importing/StoreImportSourceImpl.java
@@ -34,7 +34,10 @@ public class StoreImportSourceImpl extends StoreImportSource implements CommandI
 		validateCommand(em);
 		Result result = new Result();
 		if (removeImportSource) {
-			em.remove(importSource);
+			if (importSource.getId() != null) {
+				importSource = em.find(ImportSource.class, importSource.getId());
+				em.remove(importSource);
+			}
 		} else {
 
 			// store changes to import source
